appdb refactoring + impl of headers/burnstate dbs
